Created <Authenticator /> higher order component, Renamed <LoggedInHeader /> --> <Header />, Removed <LoggedIn /> component(no longer in use)
