--- src/frontend/src/components/LoggedIn/index.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import LoggedInHeader from '../LoggedInHeader';
-
-class LoggedIn extends React.Component {
-  public render() {
-    return (
-      <div className="center">
-        <LoggedInHeader />
-        {this.props.children}
-      </div>
-    );
-  }
-}
-
-export default LoggedIn;
